+++ Changelog.md
@@ -1,3 +1,9 @@
 ## Ignition Sensors
 
+### Ignition Sensors 2.X.X (201X-XX-XX)
+
+1. Add magnetometer
+    * [Pull request ](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/)
+
+
 ### Ignition Sensors 1.0.0 (2019-03-01)